Add indicators to Register and Login, refactor indicator to separate fun
