+++ app/src/main/java/com/realityexpander/tasky/auth_feature/presentation/login_screen/LoginScreen.kt
@@ -29,6 +29,7 @@ import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.RootNavGraph
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
+import com.realityexpander.observeconnectivity.IInternetConnectivityObserver
 import com.realityexpander.tasky.MainActivity
 import com.realityexpander.tasky.R
 import com.realityexpander.tasky.auth_feature.presentation.components.EmailField
@@ -36,6 +37,7 @@ import com.realityexpander.tasky.auth_feature.presentation.components.PasswordFi
 import com.realityexpander.tasky.core.data.settings.saveAuthInfo
 import com.realityexpander.tasky.core.presentation.common.modifiers.*
 import com.realityexpander.tasky.core.presentation.theme.TaskyTheme
+import com.realityexpander.tasky.core.util.InternetConnectivityObserver.ShowInternetAvailabilityIndicator
 import com.realityexpander.tasky.dataStore
 import com.realityexpander.tasky.destinations.AgendaScreenDestination
 import com.realityexpander.tasky.destinations.RegisterScreenDestination
@@ -55,6 +57,9 @@ fun LoginScreen(
     viewModel: LoginViewModel = hiltViewModel(),
 ) {
     val loginState by viewModel.loginState.collectAsState()
+    val connectivityState by viewModel.onlineState.collectAsState(
+        initial = IInternetConnectivityObserver.OnlineStatus.OFFLINE // must start as Offline
+    )
 
     LoginScreenContent(
         username = username,                // passed to/from RegisterScreen (not used in LoginScreen)
@@ -63,6 +68,8 @@ fun LoginScreen(
         onAction = viewModel::sendEvent,
         navigator = navigator,
     )
+
+    ShowInternetAvailabilityIndicator(connectivityState)
 }
 
 @Composable
