+++ app/src/main/java/com/realityexpander/tasky/auth_feature/presentation/register_screen/RegisterScreen.kt
@@ -29,6 +29,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
+import com.realityexpander.observeconnectivity.IInternetConnectivityObserver
 import com.realityexpander.tasky.R
 import com.realityexpander.tasky.auth_feature.presentation.components.EmailField
 import com.realityexpander.tasky.auth_feature.presentation.components.NameField
@@ -36,6 +37,7 @@ import com.realityexpander.tasky.auth_feature.presentation.components.PasswordFi
 import com.realityexpander.tasky.core.presentation.common.modifiers.*
 import com.realityexpander.tasky.core.presentation.theme.TaskyShapes
 import com.realityexpander.tasky.core.presentation.theme.TaskyTheme
+import com.realityexpander.tasky.core.util.InternetConnectivityObserver.ShowInternetAvailabilityIndicator
 import com.realityexpander.tasky.destinations.LoginScreenDestination
 
 @Composable
@@ -53,12 +55,17 @@ fun RegisterScreen(
     viewModel: RegisterViewModel = hiltViewModel(),
 ) {
     val registerState by viewModel.registerState.collectAsState()
+    val connectivityState by viewModel.onlineState.collectAsState(
+        initial = IInternetConnectivityObserver.OnlineStatus.OFFLINE // must start as Offline
+    )
 
     RegisterScreenContent(
         state = registerState,
         onAction = viewModel::sendEvent,
         navigator = navigator,
     )
+
+    ShowInternetAvailabilityIndicator(connectivityState)
 }
 
 @OptIn(ExperimentalMaterialApi::class)
