+++ app/src/main/java/com/realityexpander/tasky/di/InternetConnectivityObserver.kt
@@ -2,7 +2,7 @@ package com.realityexpander.tasky.di
 
 import android.content.Context
 import com.realityexpander.observeconnectivity.IInternetConnectivityObserver
-import com.realityexpander.tasky.core.util.ConnectivityObserver.InternetConnectivityObserverImpl
+import com.realityexpander.tasky.core.util.InternetConnectivityObserver.InternetConnectivityObserverImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
